requirements.txt
# Core
google-genai>=0.3.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pydantic>=2.0.0
pyyaml>=6.0.0
google-api-core>=2.28.1

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Orchestration (optional)
# prefect>=2.0.0
